+++ sdk/storage/azure-storage-file-datalake/src/main/java/com/azure/storage/file/datalake/implementation/ServicesImpl.java
@@ -16,10 +16,12 @@ import com.azure.core.annotation.ServiceMethod;
 import com.azure.core.annotation.UnexpectedResponseExceptionType;
 import com.azure.core.http.rest.PagedResponse;
 import com.azure.core.http.rest.PagedResponseBase;
+import com.azure.core.http.rest.ResponseBase;
 import com.azure.core.http.rest.RestProxy;
 import com.azure.core.util.Context;
 import com.azure.storage.file.datalake.implementation.models.FileSystem;
-import com.azure.storage.file.datalake.implementation.models.ServicesListFileSystemsResponse;
+import com.azure.storage.file.datalake.implementation.models.FileSystemList;
+import com.azure.storage.file.datalake.implementation.models.ServicesListFileSystemsHeaders;
 import com.azure.storage.file.datalake.models.DataLakeStorageException;
 import reactor.core.publisher.Mono;
 
@@ -51,7 +53,7 @@ public final class ServicesImpl {
         @Get("/")
         @ExpectedResponses({200})
         @UnexpectedResponseExceptionType(DataLakeStorageException.class)
-        Mono<ServicesListFileSystemsResponse> listFileSystems(
+        Mono<ResponseBase<ServicesListFileSystemsHeaders, FileSystemList>> listFileSystems(
                 @HostParam("url") String url,
                 @QueryParam("resource") String resource,
                 @QueryParam("prefix") String prefix,
@@ -114,6 +116,6 @@ public final class ServicesImpl {
                                         res.getHeaders(),
                                         res.getValue().getFilesystems(),
                                         null,
-                                        res.getDeserializedHeaders()));
+                                        null));
     }
 }