+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -3,8 +3,10 @@
 ## 12.10.0-beta.2 (Unreleased)
 
 ### Features Added
-
-### Breaking Changes
+- Added ability to set permission, umask, owner, group, ACL, lease, and expiry time to the following:
+  DataLakePathClient.createWithResponse, DataLakeDirectoryClient.createFileWithResponse,
+  DataLakeDirectoryClient.createSubdirectoryWithResponse DataLakeFile.createWithResponse,
+  DataLakeFileSystemClient.createFileWithResponse, and DataLakeFileSystemClient.createDirectoryWithResponse.
 
 ### Bugs Fixed